Fix go build issue for fou code
