+++ fou_test.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package netlink
 
 import (
